backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
openai==1.61.0
feedparser==6.0.11
httpx==0.28.1
pyyaml==6.0.2
python-multipart==0.0.20
